Fixing some failed feature (#98)
* fix

* fix mock

+++ src/main/scala/com/seanmcapp/repository/dota/HeroAttributeRepo.scala
@@ -8,7 +8,7 @@ import scala.concurrent.Future
 import scala.util.{Failure, Success}
 
 case class HeroAttribute(id: Int, base_health: Int, base_health_regen: Double, base_mana: Int, base_mana_regen: Double,
-                         base_armor: Int, base_mr: Int, base_attack_min: Int, base_attack_max: Int, base_str: Int, base_agi: Int, base_int: Int,
+                         base_armor: Double, base_mr: Int, base_attack_min: Int, base_attack_max: Int, base_str: Int, base_agi: Int, base_int: Int,
                          str_gain: Double, agi_gain: Double, int_gain: Double, attack_range: Int, projectile_speed: Int, attack_rate: Double,
                          move_speed: Int, turn_rate: Double, cm_enabled: Boolean)
 
@@ -28,7 +28,7 @@ class HeroAttributeInfo(tag: Tag) extends Table[HeroAttribute](tag, "hero_attrib
   val baseHealthRegen = column[Double]("base_health_regen")
   val baseMana = column[Int]("base_mana")
   val baseManaRegen = column[Double]("base_mana_regen")
-  val baseArmor = column[Int]("base_armor")
+  val baseArmor = column[Double]("base_armor") // in postgres is int
   val baseMR = column[Int]("base_mr")
   val baseAttackMin = column[Int]("base_attack_min")
   val baseAttackMax = column[Int]("base_attack_max")
